Golem Model Implemented
The Golem is rigged to animate properly. However, it's collision is not yet moving with it's animations. Going to maybe have to move the collision meshes as part of the animations themselves, or figure out another way to have it update itself. Regardless, I'm working on it! At least it *looks* a little better!